+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 ### Fixed
 - Added improved error handling around unexpected `media_type` values
 - Improve handling of NYPL Catalog records without EDD links
-- Future government document works in ES now have 'True' boolean value in the is_government_document field
+- Future government document works in ES now have 'True' boolean value in their is_government_document field
 
 ## 2022-06-09 -- v0.10.4
 ### Added